+++ src/main/java/pl/asie/computronics/api/audio/IAudioReceiver.java
@@ -4,7 +4,7 @@ import net.minecraft.world.World;
 
 import net.minecraftforge.common.util.ForgeDirection;
 
-public interface IAudioReceiver {
+public interface IAudioReceiver extends IAudioConnection {
 	World getSoundWorld();
 	int getSoundX();
 	int getSoundY();