+++ src/main/java/pl/asie/computronics/api/audio/IAudioSource.java
@@ -1,5 +1,5 @@
 package pl.asie.computronics.api.audio;
 
-public interface IAudioSource {
+public interface IAudioSource extends IAudioConnection {
 	int getSourceId();
 }