+++ changelog.md
@@ -4,10 +4,13 @@
     - Connects audio players (like Tape Drives) to audio receivers (Like Speakers)
     - If you connect one to a Tape Drive, it will stop playing itself and instead send the sound across the cables.
     - Can be coloured. When coloured, they won't connect to cables with a different colour and won't connect to coloured audio senders/receivers with a different colour.
+    - They also support Immibis' Microblocks for separation, if you like covering up your cabling.
   * ADDED: Speakers!
     - Can be connected to Tape Drives or Cables
     - Plays any sound it receives through the cables
     - Allows you to play sound in multiple places at once using only a single Tape Drive.
+  * ADDED: Audio API!
+    - Now you can create your own audio packets! Sort of - it's a bit rough around the edges.
   * CHANGED: NedoComputers is no longer supported. Removed the EEPROM Reader and any kind of NedoComputers support there was.
   * FIXED: More crashes related to invalid crafting recipes.